Refactor scanner to use TokenInfo for token representation; enhance error handling and update main script for file processing
